Package the application into Dongji package
The application was in default directory, and not packaged.

Create new package Dongji and move the entire project into it.

Let's move the application into valid package.
