+++ src/main/java/Dongji/components/commands/Command.java
@@ -1,0 +1,7 @@
+package Dongji.components.commands;
+
+
+public interface Command {
+
+    public String execute();
+}
