+++ src/main/java/Dongji/components/commands/ListCommand.java
@@ -1,0 +1,16 @@
+package Dongji.components.commands;
+
+import Dongji.components.tasks.TaskList;
+
+public class ListCommand implements Command {
+    private TaskList taskList;
+
+    public ListCommand(TaskList taskList) {
+        this.taskList = taskList;
+    }
+
+    @Override
+    public String execute() {
+        return this.taskList.listTasks();
+    }
+}
