+++ src/main/java/Dongji/components/parsers/DateTimeParser.java
@@ -1,0 +1,17 @@
+package Dongji.components.parsers;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
+import Dongji.components.DateTimeData;
+
+public class DateTimeParser {
+    public static DateTimeData extractDateTime(String input) throws DateTimeParseException{
+        boolean hasTime = input.contains(" ");
+        if (! hasTime) {
+            input += " 0000";
+        }
+        return new DateTimeData(LocalDateTime.parse(input, DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm")), hasTime);
+    }
+}
