+++ src/main/java/Dongji/components/persistences/Persistence.java
@@ -1,0 +1,9 @@
+package Dongji.components.persistences;
+
+import Dongji.components.tasks.TaskList;
+
+public interface Persistence {
+    public void exportTasks();
+
+    public TaskList importTasks();
+}
