+++ src/main/java/Dongji/components/tasks/Todo.java
@@ -1,0 +1,13 @@
+package Dongji.components.tasks;
+
+import Dongji.exceptions.DongjiEmptyTaskNameException;
+
+public class Todo extends Task {
+    public Todo(String name) throws DongjiEmptyTaskNameException {
+        super(name);
+    }
+
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
