+++ src/main/java/Dongji/components/ui/Ui.java
@@ -1,0 +1,10 @@
+package Dongji.components.ui;
+
+public interface Ui {
+
+    public void show(String str);
+
+    public void showWelcome();
+
+    public void showGoodbye();
+}
