+++ src/main/java/Dongji/exceptions/DongjiEmptyTaskNameException.java
@@ -1,0 +1,7 @@
+package Dongji.exceptions;
+
+public class DongjiEmptyTaskNameException extends DongjiException {
+    public DongjiEmptyTaskNameException(String message) {
+        super(message);
+    }
+}
