+++ src/main/java/Dongji/exceptions/DongjiException.java
@@ -1,0 +1,7 @@
+package Dongji.exceptions;
+
+public class DongjiException extends Exception {
+    public DongjiException(String message) {
+        super("OOPS!!! " + message);
+    }
+}
