+++ src/main/java/Dongji/exceptions/DongjiIndexOutOfBoundException.java
@@ -1,0 +1,7 @@
+package Dongji.exceptions;
+
+public class DongjiIndexOutOfBoundException extends DongjiException {
+    public DongjiIndexOutOfBoundException(String message) {
+        super(message);
+    }
+}
