+++ src/main/java/Dongji/exceptions/DongjiParseException.java
@@ -1,0 +1,7 @@
+package Dongji.exceptions;
+
+public class DongjiParseException extends DongjiException {
+    public DongjiParseException(String message) {
+        super(message);
+    }
+}
