+++ src/main/java/Dongji/exceptions/DongjiUnknownInstructionException.java
@@ -1,0 +1,7 @@
+package Dongji.exceptions;
+
+public class DongjiUnknownInstructionException extends DongjiException {
+    public DongjiUnknownInstructionException(String message) {
+        super(message);
+    }
+}
