--- src/main/java/components/commands/Command.java
@@ -1,7 +1,0 @@
-package components.commands;
-
-
-public interface Command {
-
-    public String execute();
-}
