--- src/main/java/components/commands/ListCommand.java
@@ -1,16 +1,0 @@
-package components.commands;
-
-import components.tasks.TaskList;
-
-public class ListCommand implements Command {
-    private TaskList taskList;
-
-    public ListCommand(TaskList taskList) {
-        this.taskList = taskList;
-    }
-
-    @Override
-    public String execute() {
-        return this.taskList.listTasks();
-    }
-}
