--- src/main/java/components/parsers/DateTimeParser.java
@@ -1,17 +1,0 @@
-package components.parsers;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-
-import components.DateTimeData;
-
-public class DateTimeParser {
-    public static DateTimeData extractDateTime(String input) throws DateTimeParseException{
-        boolean hasTime = input.contains(" ");
-        if (! hasTime) {
-            input += " 0000";
-        }
-        return new DateTimeData(LocalDateTime.parse(input, DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm")), hasTime);
-    }
-}
