--- src/main/java/components/persistences/Persistence.java
@@ -1,9 +1,0 @@
-package components.persistences;
-
-import components.tasks.TaskList;
-
-public interface Persistence {
-    public void exportTasks();
-
-    public TaskList importTasks();
-}
