--- src/main/java/components/tasks/Todo.java
@@ -1,13 +1,0 @@
-package components.tasks;
-
-import exceptions.DongjiEmptyTaskNameException;
-
-public class Todo extends Task {
-    public Todo(String name) throws DongjiEmptyTaskNameException {
-        super(name);
-    }
-
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
