--- src/main/java/components/ui/Ui.java
@@ -1,10 +1,0 @@
-package components.ui;
-
-public interface Ui {
-
-    public void show(String str);
-
-    public void showWelcome();
-
-    public void showGoodbye();
-}
