--- src/main/java/exceptions/DongjiEmptyTaskNameException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DongjiEmptyTaskNameException extends DongjiException {
-    public DongjiEmptyTaskNameException(String message) {
-        super(message);
-    }
-}
