--- src/main/java/exceptions/DongjiException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DongjiException extends Exception {
-    public DongjiException(String message) {
-        super("OOPS!!! " + message);
-    }
-}
