--- src/main/java/exceptions/DongjiIndexOutOfBoundException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DongjiIndexOutOfBoundException extends DongjiException {
-    public DongjiIndexOutOfBoundException(String message) {
-        super(message);
-    }
-}
