--- src/main/java/exceptions/DongjiParseException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DongjiParseException extends DongjiException {
-    public DongjiParseException(String message) {
-        super(message);
-    }
-}
