--- src/main/java/exceptions/DongjiUnknownInstructionException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DongjiUnknownInstructionException extends DongjiException {
-    public DongjiUnknownInstructionException(String message) {
-        super(message);
-    }
-}
